+++ HISTORY.rst
@@ -8,3 +8,5 @@ History
 * Added new parameter **networkattributes** for POST endpoint that
   lets caller set network attributes for base network uploaded to NDEx
 
+* Added new endpoint /<task id>/rawclusteringoutput that contains
+  raw output from clustering algorithm used